Renaming ConnectTheDotsCloudDeploy solution to Azure Prep
Renaming process to make it more in line with the task
