+++ Azure/AzurePrep/AzurePrep.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
 VisualStudioVersion = 12.0.31101.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConnectTheDotsCloudDeploy", "ConnectTheDotsCloudDeploy\ConnectTheDotsCloudDeploy.csproj", "{5A2497C7-09B1-4C66-9D45-7F68F4D7628F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AzurePrep", "AzurePrep\AzurePrep.csproj", "{5A2497C7-09B1-4C66-9D45-7F68F4D7628F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
