+++ Azure/WebSite/ConnectTheDotsWebSite/Web.config
@@ -32,12 +32,12 @@ http://msdn2.microsoft.com/en-us/library/b5ysx397.aspx
   </system.web>
   <appSettings>
 
-    <add key="Microsoft.ServiceBus.EventHubDevices" value="ehdevices" />
-    <add key="Microsoft.ServiceBus.EventHubAlerts" value="ehalerts" />
-    <add key="Microsoft.ServiceBus.ConnectionString" value="Endpoint=sb://{namespace-name}.servicebus.windows.net/;SharedSecretIssuer=owner;SharedSecretValue={key}" />
-    <add key="Microsoft.ServiceBus.ConnectionStringDevices" value="Endpoint=sb://{namespace-name}.servicebus.windows.net/;SharedAccessKeyName=WebSite;SharedAccessKey={key}" />
-    <add key="Microsoft.ServiceBus.ConnectionStringAlerts" value="Endpoint=sb://{namespace-name}.servicebus.windows.net/;SharedAccessKeyName=WebSite;SharedAccessKey={key}" />
-    <add key="Microsoft.Storage.ConnectionString" value="DefaultEndpointsProtocol=https;AccountName={storageaccount};AccountKey={key}" />
+    <add key="Microsoft.ServiceBus.EventHubDevices" value="EHDevices" />
+    <add key="Microsoft.ServiceBus.EventHubAlerts" value="EHAlerts" />
+    <add key="Microsoft.ServiceBus.ConnectionString" value="Endpoint=sb://feb23ctd-ns.servicebus.windows.net/;SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=/yR9wzVRgGoT4cZNarzenl9wdci2PIJ/KYir81vjlZo=" />
+    <add key="Microsoft.ServiceBus.ConnectionStringDevices" value="Endpoint=sb://feb23ctd-ns.servicebus.windows.net/;SharedAccessKeyName=WebSite;SharedAccessKey=2syDJzBkUD89i3k2jSzhaDHScKS0hSmejYMmQbiTapc=" />
+    <add key="Microsoft.ServiceBus.ConnectionStringAlerts" value="Endpoint=sb://feb23ctd-ns.servicebus.windows.net/;SharedAccessKeyName=WebSite;SharedAccessKey=K02ZtJKNA+krrRNLXHg5+PeZIsJaZ5MYQxkExeStE9A=" />
+    <add key="Microsoft.Storage.ConnectionString" value="DefaultEndpointsProtocol=https;AccountName=feb23ctdstorage;AccountKey=uPY6Yvh8KkhyHmyRD55/8xl5F0IQ65JXR/Iweab8IVan3axROmdVQD/hct338MuzpLUKgCogs8oJ7XSU6BZKXw==" />
 
     <add key="aspnet:UseTaskFriendlySynchronizationContext" value="true" />
   </appSettings>
